Removed submitMovingFundsCommitment from MaintainerProxy
The Bridge requires `submitMovingFundsCommitment` to be called by one of the
wallet members, this function can not be called and reimbursed via
`MaintainerProxy`.

The implementation in `MaintainerProxy.submitMovingFundsCommitment` was
incorrect and the tests were passing only because they had an incorrect
assumption of `MaintainerProxy` contract being the wallet member:
```
walletRegistry.isWalletMember
  .whenCalledWith(
    ecdsaWalletTestData.walletID,
    walletMembersIDs,
    maintainerProxy.address,
    walletMemberIndex
  )
.returns(true)
```

`MaintainerProxy` contract is not an operator and can not be a wallet member.