Delete entry from `timedOutRedemptions` when it has been processed
In `BridgeState` we used to say that:
> Timed out requests are stored in this mapping to avoid slashing the wallets
> multiple times for the same timeout.

This was not true, because when it comes to requesting redemption we say that:
> There is no need to check for existence in timedOutRedemptions since the
> wallet's state is changed to other than Live after first time out is reported
> so making new requests is not possible.

And when the redemption timeout is reported we say that:
> It is worth noting that there is no need to check if timedOutRedemption
> mapping already contains the given redemption key. There is no possibility
> to re-use a key of a reported timed-out redemption because the wallet
> responsible for causing the timeout is moved to a state that prevents it to
> receive new redemption requests.

Given that the wallet lifecycle prevents the wallet from being slashed twice for
the redemption timeout, we can delete an entry from `timedOutRedemptions`
mapping when the timed-out redemption has been processed.

Redemption request is deleted from `pendingRedemptions` when it is processed.
Pending redemption can be processed in two ways: it can be used in SPV proof and
it can be marked as timed-out. Deleting `timedOutRedemptions` when it is
processed is consistent with how `pendingRedemptions` are deleted and leaves no
dangling state in the Bridge. A timed-out redemption in `timedOutRedemptions`
can be processed in just one way: used in SPV proof.