Use `development` tag for `tbtc` dependency
The range that we had in `package.json` for `@keep-network/tbtc`
dependency was no longer containing the packages that we wanted. Since
we introduced `-goerli` suffixed packages, those packages were also
withing the range (and we want only `-dev` suffixed packages to be
within the range). Instead of using suffixes in the ranges we can also
use tags, which should be easier to maintain. So now we use
`development` tag, which should always point to the latest `tbtc`
package meant for development purposes.