Use forked bcoin repo to fix browser issue
We are using the forked bcoin repo - keep-network/bcoin - to fix dependency of
bcoin (bcrypto) that is not properly mapped in package.json under "browser"
property. The commit hash that we are using is
5accd32c63e6025a0d35d67739c4a6e84095a1f8 (it's from keep-network/bcoin fork).

Thanks to that the bcoin lib should work properly in the browser when creating
`new bcoin.Script()`.