EIP-2612 complicance fix: n/nonce/nonces
The Bank smart contract implements EIP-2612 but used mapping `nonce` (singular)
instead of `nonces`. This might create an issue when third-party libraries
(that support EIP-2612) try to interact with the Bank and assume that nonces are
tracked with `nonces` instead of `nonce`. Fixed it by renaming `nonce` mapping
to `nonces`.