Warn about 32-bits timestamp representation limitation
The timestamp/locktime is represented using 32 bits for gas efficiency and to be
consistent with Bitcoin.

This way, OperatorRewards structure in Rewards contract takes a single EVM word
and the reward withdrawal consumes less gas. DepositRequest structure in Deposit
library of the Bridge contract takes two EVM words and the operation revealing
a deposit and the operation sweeping a deposit consume less gas. Wallet
structure takes 4 EVM words and wallet lifecycle operations consume less gas.

FraudChallenge, MovingFundsSweepRequest ,Redemtion libraries of Bridge contract
represent the timestamp as uint32 for consistency with the rest of the code.

Note that Bitcoin represents the locktime using 32 bits as well:
https://developer.bitcoin.org/devguide/transactions.html#locktime-and-sequence-number

Added a comment to the code to make it clear the uint32 representation is
sufficient only until 2106.