RFC-1: Adjusted sweep frequency based on RFC-7 findings
The time it takes to sign a sweep transaction input and Ethereum block gas
limits put a cap on how many deposits can be swept at once. It makes more sense
to trigger sweep early, given the number of deposits rather than look at the
amount of BTC deposited because it does not affect the submit SPV proof
transaction that much.