Optimistic minting early draft
This is a very draft implementation presenting the idea. Tests are failing, some
functionality is missing, and the code may not work as expected.

There are two permissioned sets in the system: Minters and Watchmen, both set
up in 1-of-n mode. Minters observe the revealed deposits and request minting
TBTC. Any single Minter can perform this action. There is a 2-3 hours delay
between the time of the request from a Minter to the time TBTC is minted. During
the time of the delay, any Watchmen can cancel the minting.

So far we kept vaults completely separate from the Bridge. The Bridge was
managing Bank balances and vaults - including TBTCVault - were just applications
working on a top of the Bank. Neither Bank nor the vaults were interested in how
the Bridge worked. They knew nothing about revealed deposits and sweeping.

Given that Minter should be only able to request minting for a revealed, unswept
deposit, this separation may no longer be kept. At the same time, the goal is to
minimize the impact and keep this "abstraction breaking change" local to the
optimistic minting component the TBTCVault will extend. This will let us easily
unplug this mechanism in the future, if needed.

For any revealed, unswept deposit, Minter can request minting using
`function optimisticMint(uint256 depositKey) external onlyMinter` on `TBTCVault`.

Watchman can cancel the minting by calling
`function cancelOptimisticMint(uint256 depositKey) external onlyWatchman`.

If none of the watchmen canceled minting, any Minter can finalize the request
calling `function finalizeOptimisticMint(uint256 depositKey) external onlyMinter`
to mint TBTC. When TBTC is minted, depositor's optimistic minting debt balance
is updated in `optimisticMintingDebt`. When a deposit is swept for the depositor,
the vault needs to update `optimisticMintingDebt` and if this amount is non-zero,
pay back the debt and mint only the amount of TBTC that is left after paying the
debt.

This approach allows us to avoid passing deposit ID from the `Bridge`, through
`Bank`, to `TBTCVault` which would break the abstractions and require a lot of
changes in a quite complicated, gas-optimized `Bridge` contract. It also allows
us to track - looking only at Ethereum - the amount of BTC that is in the queue
for sweeping and got optimistically minted into ERC20 already.