Deduct treasury deposit fee when optimistically minting TBTC
Bridge, when sweeping, cuts a deposit treasury fee and splits Bitcoin miner fee
for the sweep transaction evenly between the  depositors in the sweep.

When tokens are optimistically minted, we do not know what the Bitcoin miner fee
for the sweep transaction will look like. The Bitcoin miner fee is ignored. When
sweeping, the miner fee is  subtracted so the optimisticMintingDebt may stay
non-zero after the  deposit is swept.

This imbalance is supposed to be solved by a donation to the Bridge.