Listed validations that the Guardians must perform
The following validations have to be performed by Guardians:
- The deposit happened on Bitcoin side and it has enough confirmations.
- The optimistic minting has been requested early enough so that the wallet has
  enough time to sweep the deposit.
- The wallet is an active one and it does perform sweeps or it will perform
  sweeps once the sweeps are activated.

Given that the Guardians validate the locktime, I removed the TODO from the
contract. This is better for gas efficiency: the contract would have to perform
the same validation as `validateDepositRefundLocktime` and expect the entire
`DepositRevealInfo` to be passed to assemble the expected script hash on-chain.