Refer to bridge field from TBTCVault instead of going via bank.bridge()
We used `bank.bridge()` in `TBTCVault` to have a single update point in case
replacing the entire Bridge is needed. Before optimistic minting, `TBTCVault`
worked only with `Bank` and was not interested in how the `Bridge` worked.
Hence, it was possible to do `bank.bridge()` to obtain the `Bridge` address for
the redemption request. Nothing more was needed. With the optimistic minting,
`TBTCVault` must inspect the deposit state, so updating Bridge is no longer that
easy. That being said, we can refer to the `bridge` field from
`TBTCOptimisticMinting` the `TBTCVault` inherits from.