Accept zero Bridge redemption treasury fee
We have a condition in `Redemption.sol` in the `requestRedemption` function
allowing to properly handle the case when the Bridge's redemption treasury fee
is zero:

```
uint64 treasuryFee = self.redemptionTreasuryFeeDivisor > 0
    ? amount / self.redemptionTreasuryFeeDivisor
    : 0;
```

At the same time, we were not allowing for a zero `redemptionTreasuryFeeDivisor`
parameter in the validation of update parameter value functions.

This has been fixed and zero `redemptionTreasuryFeeDivisor` is allowed now.
I have also added unit tests ensuring requesting redemption with a zero
treasury fee works fine. The case of providing a redemption proof when the fee
is zero is already covered, see all the `bridge.setRedemptionTreasuryFeeDivisor(0)`
calls across `Bridge.Redemption.test.ts`.