Made it clear how the optimistic minting fee is computed
We could apply the optimistic minting fee based on the original deposit amount,
but that would lead to inconsistent amounts if the optimistic minting fee would
be high. It is easiest to explain it with an example. Say that the deposit
amount is 1000 and that the Bridge fee is 20. 980 is what lands in TBTCVault. If
the optimistic deposit fee was anything > 98%, we would be printing tokens
because only 980 out of the original 1000 lands in the vault. In other words,
the vault can cut a fee only from what it is actually receiving. We could
probably introduce some kind of validation between fees but that would break the
separation between vaults (applications) and the Bridge.