Allow only the depositor to reveal a deposit
Consider this scenario:
- Alice reveals her Bitcoin deposit and routes it to vault 0xABC
- Eve notices Alice's TX in the mempool and frontruns it with the same data but
  the vault 0xDEF

As a result, Alice's deposit lands in vault 0xDEF and not in vault 0xABC.
Both 0xDEF and 0xABC have to be trusted vaults but it is still an issue given it
is not what Alice was asking for.

Here we change the validation requirements of Bridge.revealDeposit function.
From now on, this function must be called by the same Ethereum address as the
one used in the P2(W)SH BTC deposit transaction as a depositor.