Use hardhat node --no-deploy
Due to how `hardhat-deploy` plugin works we need to use `--no-deploy`
option.

We expect that system tests are executed against `localhost` network.
This network name is also used in the relay deployment script to deploy
TestRelay contract instead of LightRelay.

`hardhat-deploy` plugin modifies `hardhat node` task to run deployment
scripts along with the node startup. This causes a problem, as when node
starts the `hardhat` network is used instead of `localhost`.
We need to separate node running from deployment execution, so here
comes a change from this commit.