Optimistic minting debt calculation fix
The optimistic debt is explained in the documentation in the following way:

The debt represents the total value of all depositor's deposits revealed to the
Bridge that has not been yet swept and led to the optimistic minting of TBTC.
(...)

This explanation is perfectly correct but it is not how the debt was calculated
for a non-zero optimistic minting fee. Previously, the optimistic minting fee
was deducted from the debt. This is not correct - the fee is also optimistically
minted and goes to the treasury address. All TBTC that is optimistically minted:
the part that goes to the depositor and the part that goes to the treasury
should be added to the optimistic minting debt. When the deposit is swept, it
is paying off both the depositor's part and the treasury's part.