Exclude typechain script from postinstall
We use typechain script to generate TS bindings for contracts that we
install as dependencies. With postinall script the bindings were
generated always when we installed the dependencies.
Unfortunatelly if this package was being used in another project as
dependency this postinstall script was also executed, which doesn't make
sense as we publish the generated bindings as part of the package.

With this change here we move the typechain bindings geneartion to the
build script so they are generated when building the package before
publication.