Fix the TBTC token ownership update in deployment scripts for Goerli
Unit tests cover the path of the VendingMachine upgrade to TBTCVault. We
transfer the ownership of TBTC token to the VendingMachine for Hardhat network
used for unit tests to enable this path. For all other environments (e.g. local
dev, Goerli), we want the TBTCVault to be the owner of TBTC token. This
simulates the final 1.0 version mainnet state.