Define scriptSig and scriptPubKey as Hex
Before a7aedd11fbfc7b26cbda5376a33207c7966f48bf
under these two properties we returned exactly the valu that was
retruned from electrum server. This was causing some unexpected
bahaviour as it could happen that different implementation and instances
of servers were returning different results.
The only thing that we can relay on is a hex value for bytes that were
used in the transaction.
Here we switch to hexadecimal value to have consistent behaviour and be
able to support all major electrum server implementations (ElectrumX,
Electrs/Esplora and Fulcrum).
The change here is similiar to what we implemented in the Keep Client.