Make `decodeBitcoinAddress` throw on non-PKH addresses
The `decodeBitcoinAddress` documentation says that this function decodes
P2PKH and P2WPKH addresses into a public key hash. However, this requirement
is not enforced by the function code. Here we add a validation that makes
the function throw if the passed address is not PKH-based. That means the
function will throw an exception if the passed address is, for example, P2SH or
P2WSH.