Explicitly rebuild bcrypto and use node 14
In 0e1f62414b5c804c690fd32d4270b8de8a5d4cb3 we added a requirement to
rebuild on postinstall. In this commit we make it explicit that we need
to rebuild bcrypto.

Also the rebuild works only on node 14, which is confusing (see:
https://github.com/keep-network/tbtc-v2/issues/509) so we update engines
to be explicit that node 14 is required.