Make the monitoring k8s-ready
So far, the plan was to deploy the monitoring as a Cloud Function triggered
using the Cloud Scheduler. However, the function became quite heavy and
does no longer fit well into the Cloud Functions environment. Here we pivot
to a k8s deployment by making necessary adjustments. The most important is
switching from GCS-bucket-based storage to a simple file that can be mounted
to a k8s pod. We are also getting rid of unnecessary dependencies.