L2WormholeGateway: send and receive tokens
This is a draft implementation of the L2WormholeGateway smart contract.

To be done:
- unit tests
- events
- an absolute, governable minting limit for early days (?)

Selected cross-ecosystem bridges are given the minting authority for tBTC token
on L2 and sidechains. This contract gives a minting authority to the Wormhole
Bridge.

The process of bridging from L1 to L2 (or sidechain) looks as follows:
1. There is a tBTC holder on L1. The holder goes to the Wormhole Portal and
   selects the chain they want to bridge to.
2. The holder submits one transaction to L1 locking their tBTC tokens in the
   bridge’s smart contract. After the transaction is mined, they wait about 15
   minutes for the Ethereum block finality.
3. The holder submits one transaction to L2 that is minting tokens. After that
   transaction is mined, they have their tBTC on L2.

The process of bridging from L2 (or sidechain) to L1 looks as follows:
1. There is a tBTC holder on L2. That holder goes to the Wormhole Portal and
   selects one of the L2 chains they want to bridge from.
2. The holder submits one transaction to L2 that is burning the token. After the
   transaction is mined, they wait about 15 minutes for the L2 block finality.
3. The holder submits one transaction to L1 unlocking their tBTC tokens from the
   bridge’s smart contract. After that transaction is mined, they have their
   tBTC on L1.

This smart contract is integrated with step 3 of L1->L2 bridging and step 1 of
L2->L1 or L2->L2 bridging. When the user redeems token on L2, this contract
receives the Wormhole tBTC representation and mints the canonical tBTC in an
equal amount. When user sends their token from L1, this contract burns the
canonical tBTC and sends Wormhole tBTC representation through the bridge in an
equal amount.

This contract is supposed to be deployed behind a transparent upgradeable proxy.