L2WormholeGateway: send tokens to chain with no Wormhole tBTC gateway
We had a missing path in L2WormholeGateway where tokens were moved between L2s
and the target L2 had Wormhole tBTC gateway deployed capable of minting
canonical tBTC. In this case, the transfer should be done with a payload
allowing the user to receive tBTC on the target chain using the gateway
deployed there.

The previous approach was fine when tBTC was transferred back to L1.
It was not enough when tokens were transferred between L2s having canonical
token minted by gateways.

The missing path has been covered in the sendTbtc function. Depending whether
there is a gateway on the target chain, transfer with payload or transfer with
no payload is used.