Adding Cross-chain Optimism module
This commit is a continuation of tBTC expansion to L2s, in this case
Optimism.
In the deployment scripts we pull a generic L2TBTC from the
tbtc-v2/solidity project and then deploy it on the Optimism network.
'OptimisticTBTC' will become a canonical tBTC token with a minting
authority delegated to 'OptimisticWormholeGateway'.
'OptimisticWormholeGateway' is a smart contract wrapping and unwrapping
Wormhole-specific tBTC representation into the canonical
'OptimisticTBTC' token.

Code was largerly copied from the cross-chain/arbitrum module.

A couple of things to pay closer attention at:
- addresses under the external/ dir
- gateway address to a wormhole chain id: 24
- optimistic Goerli and Mainnet chain ids: 420 and 10