Adding cross-chain Polygon module
This commit is a continuation of tBTC expansion to L2s and sidechains.
Here we expand tBTC to Polygon.
In the deployment scripts we pull a generic L2TBTC from the
tbtc-v2/solidity project and then deploy it on the Polygon network.
'PolygonTBTC' will become a canonical tBTC token with a minting
authority delegated to 'PolygonWormholeGateway'.
'PolygonWormholeGateway' is a smart contract wrapping and unwrapping
Wormhole-specific tBTC representation into the canonical
'PolygonTBTC' token.