Add Heartbeat action to WalletCoordinator
With the existing, pretty computationally-expensive tECDSA signing algorithm,
we would prefer not to execute heartbeats at the same time as other wallet
actions (sweeping, redemptions, moving funds). Moreover, we would prefer not to
execute heartbeats from multiple wallets at the same time.

Such coordination is complicated off-chain for the same reason as the
coordination of sweeps and redemptions. We could - as we do now - execute
heartbeats at fixed hours but it does not help with coordination between
heartbeats and sweeps. Also, it does not solve the problem of multiple
heartbeats at the same time.

For this reason, the heartbeat has been made a first-class wallet action
coordinated in the on-chain contract just like the rest of wallet actions.