Run tests weekly
We're decreasing frequency of running the tests (we do not develop contracts as
actively as we used to so running them every 24h is too aggressive, especially
that the tests use testnet funds). From now on we will run tests weekly (every
Sunday at 0:00). We can still trigger the tests manually if needed.