Making a system test wait a little before a bitcoin client calls
Multiple runs of system tests showed that it happens from time to time
that a bitcoin client doesn't "sync" instantly with the bitcoin chain
after an event happens, e.g. deposit. It might not retrieve requested
data right after the call. This commit make a system test wait at
the begging of a transaction confirmation instead of at the end, which
gives some time for a bitcoin client to "sync" in the background.