Adjust WalletCoordinator gas offsets
Two gas offsets have been adjusted as a result of User Acceptance Tests
execution on Goerli.

`heartbeatRequestGasOffset` has been updated from `5_000` to `10_000`. Looking
at the heartbeat request TX
(https://goerli.etherscan.io/tx/0x0b76d88351a7e8d85fea53be4e17931d2a753ffddd90e73bcc734230c9ac84c8)
the difference between ETH burned and reimbursed is 0.00000349514053395 eth
which is 3495 Gwei. Assuming the gas price was ~1 Gwei, we should increase the
offset by ~3500 Gwei so 8_500. We add an additional 1_500 as a safety margin.

For `depositSweepProposalSubmissionGasOffset`, the situation is more complex
because the cost of the transaction depends on the number of inputs. We decided
to optimize for 10 inputs. This transaction was expected with the offset set
to `20_000` and 10 inputs:
https://goerli.etherscan.io/tx/0x3c9df2ff00b533fc4b94f905a9c5bba5c70e9594e97ec215a6cf87aa0686902a