Update `requestRedemption` fn
For testing purposes, we were passing the BTC address as
`redeemerOutputScript` and then created the output script based on this
address. Here we remove this code snipped because the output script that
the redeemed funds will be locked to is expected to be un-prefixed and
not prepended with length and we should pass the output script in a
proper form to the `requestRedemption` function.