Refactor the `requestRedemption` fn
We do not want to expose the low-level function from chain handles(see
`buildRedemptionData` from `EthereumBridge` handle). The goal of
`src/ethereum.ts` is to translate domain-level objects into low-level
Ethereum chain primitives and perform a call.

Instead of having `buildRedemptionData` in `Bridge` interface here we
modify the `approveAndCall` interface and replace `extraData: Hex` with
domain-level objects, as accepted by `buildRedemptionData`. Here we also
rename the `approveAndCall` to `requestRedemption` and explain in the
docs it happens via `approveAndCall`- although `requestRedemption` does
not exist on the Solidity code of TBTC token, having it in `TBTCToken`
Typescript API is fine given we abstract the complexity and make the API
human-readable.