Adding Base module for canonical TBTC
Most of the files were copied from the Optimism module and adjusted for
the Base chain, ie. ids and addresses. There is one change that had to
be added which is a custom network for hardhat etherscan plugin. Base is
not supported by this plugin by default and a custom network
configuration had to be added.