Basic redemptions monitoring
This changeset adds redemptions to the monitoring system. Specifically,
two new system events are supported now:

### Redemption requested

An *informational system event* indicating that a new redemption was requested
from the on-chain Bridge contract. This event is directly sent to Discord as a
notification that does not require any action.

### Large redemption requested

A *warning system event* indicating that a large redemption was requested from
the on-chain Bridge contract. This event is sent to Sentry hub and should get
team’s attention. The default action is making sure that the redemption is
not a result of a malicious action, and if not, that the redemption is
handled correctly by the system.