Stale redemptions monitoring
This changeset adds one new system event to the monitoring system:

### Stale redemption

A *warning system event* indicating that a redemption request became stale, i.e.
was not handled within the expected time. This event is sent to Sentry hub and
should get team’s attention. The default action is investigating the cause
of the extended processing time as this alert may be an early sign of
a malfunctioning wallet or may indicate a problem with the maintainer bot.