Move remaining code to the `./services/maintenance` feature module
The code left is purely related to system maintainers or is helper
tooling empowering internal operations like system testing. We
are moving it to a dedicated module.

By the way, we are arranging the main `index.ts` and adjust all imports
to use the base path and not dive into specific directories.