Make `receiveTbtc` non-reentrant
The Wormhole Token Bridge contract has protection against redeeming
the same VAA again. When a Token Bridge VAA is redeemed, its
message body hash is stored in a map. This map is used to check
whether the hash has already been set in this map. For this reason,
`receiveTbtc` does not have to be nonReentrant in theory. However,
to make this function future-proof and non-dependent on
Wormhole Bridge implementation, we are making it nonReentrant anyway.