Remove job publishing testnet `tbtc-v2.ts` packages
We no longer want to publish new `typescript` (`tbtc-v2.ts`) testnet packages as
part of the CI flow. The Typescript project is transitioning to v2, which has
all the chain configs embeded, so we'll no longer use the `mainnet` and `goerli`
packages. The only tags will be `latest` and `development`. All existing testnet
packages were marked in the NPM registry as deprecated (projects still can use
them as dependencies, but will get a warning).