Rework `README` of the typescript SDK
Here we rework the SDK `README` document by providing basic guidance for
both contributors and integrators.

Apart from that, we are changing the file type from `*.adoc` to `*.md`.
This is because only Markdown files are properly displayed in the npm registry
and. Moreover, we do not use AsciiDoc features in this file anyway.