Store `extraData` as part of `DepositRequest` struct
It turns out that emitting `extraData` as part of the existing `DepositRevealed`
event is problematic as it is not compatible backward. Adding a new field to
the event changes its signature (it is computed as keccak256 of the event name
and argument types) so off-chain clients will not be able to fetch past
`DepositRevealed` events (without the `extraData` field) using the new
signature of the `DepositRevealed` (with added `extraData` field) event.