Disambiguate the nomenclature within `WalletProposalValidator` contract
The `WalletProposalValidator` contract uses a concept of `deposit extra info`
which is an additional data required to perform validation of a deposit
sweep proposal. However, most of the comments refer that as `deposit extra data`
which is problematic in the light of the recently introduced 32-byte
deposit extra data which is a different thing.

To fix that, we are improving the applicable comments and refer to `deposit
extra info` explicitly.