Add coinbase transaction proof to SPV proof
An unusual but valid 64-byte Bitcoin transaction can be used to produce
SPV proofs for different, invalid transactions by treating the
transaction as another node in the merkle tree of the fraudulent
transaction's proof. This exploit can be prevented by checking that the
merkle proof of the coinbase transaction has the same length as the
proof of the transaction of interest.

All transactions in a valid Bitcoin block are on the same level of the
merkle tree, and it is prohibitively difficult to produce a hash
preimage that matches the first 32 bytes of a valid coinbase
transaction. Thus, if such a preimage can be provided, we can trust that
the merkle tree of the coinbase transaction is legitimate, and that the
transaction of interest is valid because it is on the same level of the
tree as the coinbase transaction.