Don't track pending deposits in AbstractTBTCDepositor contract
To optimize gas usage we removed `pendingDeposits` mapping. This
operation reduces gas usage of `_initializeDeposit` function by ~20k
gas.

After this change it is implementation's contract responsibility to
ensure `_finalizeDeposit` function cannot be called twice for the same
deposit.

`_initializeDeposit` function is not affected as the bridge blocks
possibility of revealing the same deposit twice.