Objection mechanism
Here we expose the `raiseObjection` function that allows guardians
object to specific redemption requests.

Each redemption has a default delay period during which the wallet is not
allowed to process it and the guardians can raise objections to. Each objection
extends the delay period by a certain amount of time. The third objection vetoes
the redemption request. This causes the redemption request to be rejected and
the redeemer to be penalized. Specific consequences of a veto are as follows:
- The redemption amount is frozen for a certain period of time,
- Once the freeze period expires, the redeemer can claim the frozen amount minus
  a penalty fee,
- The penalty fee is burned,
- The redeemer is banned from making future redemption requests.