Do not use Wormhole Relayer for L2 -> L1 deposit reveal
Such a deposit reveal was tested for Base -> Ethereum and turned out
to be super expensive. It takes ~0,045 ETH so ~150 USD to transfer deposit
data to Ethereum. This is an unacceptable cost. To overcome that problem,
we are abandoning the idea of using Wormhole Relay for that in favor
of our own relay bot implementation. This also makes sense as eventually,
we are aiming towards using gas-less reveals on Base.