Support for cross-chain contracts
Here we enhance the SDK to support cross-chain contracts.
We are achieving this by introducing a new `crossChainContracts`
method exposed by the main `TBTC` component. This method allows to
obtain cross-chain contracts handles for a specific supported L2 chain.

This feature will serve as a base to write a cross-chain Bitcoin depositor
component that will be used for direct bridging of deposits initiated on
supported L2 chains.

By the way, we are taking an opportunity and refactor the chain representation
used across the SDK. Namely, we are getting rid of the `EthereumNetwork`
type, and we are replacing it with a global `Chains` namespace
that aggregates all chains supported by the tBTC v2 SDK.