Authorized reimbursements in the `L1BitcoinDepositor` contract
If someone front-runs the Relayer Bot to call the `initializeDeposit` function,
the Relayer Bot's transaction will be reverted, wasting a substantial amount of
gas since the validation of duplicate deposits is performed very late in the
process, in the tBTC `Bridge` smart contract. This is cost-effective to the
front-runner, as the refund mechanism attempts to make callers slightly
net-positive.

To mitigate the above, anyone is allowed to relay deposits but reimbursement
only occurs for a closed set of relayers.