Mark L2 direct bridging as an experimental feature
The backend components of direct bridging (contracts and relayer bot) have not
been deployed yet. Despite the TypeScript SDK has been prepared to support it,
the feature is not operational and has been postponed for now. As we need to
release a new SDK version, we are marking SDK support for L2 direct bridging
as experimental until the backend is ready.