Fix `Bridge` contract artifact for Sepolia
While fetching contract events, the start block for the searched range
is taken based on the contract creation block. This is the
`receipt.blockNumber` field living in the contract artifact.
However, this field may be overwritten by the Hardhat Upgrades plugin and
point to the deployment block of the recent contract implementation, not the
transparent proxy.

This was the case for the Sepolia `Bridge` artifact. Contract
events like `NewWalletRegistered` were taken starting from the
recent implementation deployment block which occurred much later
than the `Bridge` proxy deployment. In effect, it was not possible
to fetch all `Bridge` wallets ever created.

Here we fix that problem.