Improve public key fetching for closed/terminated wallets
The current logic used to fetch wallet public keys was throwing errors while
used for `Closed` or `Terminated` wallets. This is because the `WalletRegistry`
contract does no longer hold the public key for such wallets. In effect,
the exception was propagated upstream and caused problems while searching
wallets for redemption.

Here we fix that problem by making the public key field optional and keeping
it undefined for closed/terminated wallets. That way, the client code can
detect this fact and omit closed/terminated wallets during processing.