new-deployment-of-the-L1BitcoinDepositor-on-sepolia-and-L2BitcoinDepositor-on-baseSepolia-for-base-native-minting
Deployed the updated L1BitcoinDepositor contract to the Sepolia testnet (for base native minting) and the L2BitcoinDepositor (for proper linkage to L1BitcoinDepositor). A new boolean flag, reimburseTxFee, now controls whether the txMaxFee is reimbursed to the user. This update preserves full backward compatibility by avoiding any modifications to the AbstractTBTCDepositor contract.