feat: enhance mainnet deployment configuration and add new contract addresses
### Summary

This commit updates the Hardhat configuration for the SUI project to include additional deployment paths and updates the mainnet contract addresses. It also introduces new JSON files for various contracts deployed on the mainnet, enhancing the project's cross-chain capabilities.

### Changes

- Updated `hardhat.config.ts` to include the new `suiMainnet` deployment path.
- Changed the mainnet deployer address for consistency.
- Added new deployment JSON files for `SuiBTCDepositorWormhole`, `Bridge`, `TBTCVault`, `TokenBridge`, and `Wormhole`.
- Created a new `.chainId` file to specify the mainnet chain ID.
- Introduced a new `SuiWormholeGateway` JSON file for the corresponding contract address.

These updates facilitate better management of contract deployments and improve the overall structure of the SUI integration.